Refactoring the navigation from stock Tabs to Stack for this project. Also created 2 views for Camera and a potential Picture Gallery.
