+++ src/app/index.tsx
@@ -1,0 +1,5 @@
+import HomeScreen from "@/src/app/screens/HomeScreen";
+
+export default function HomeScreenContainer() {
+  return <HomeScreen />;
+}
